requirements.txt
# Core Discord Bot Dependencies
discord.py>=2.3.0
python-dotenv>=1.0.0

# Database
aiosqlite>=0.19.0
asyncpg>=0.29.0

# Data Science and Analysis
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Natural Language Processing
nltk>=3.8.0
spacy>=3.0.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
sentence-transformers>=2.2.0
huggingface-hub>=0.19.0

# Text Processing and Matching
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.23.0
wordcloud>=1.9.0

# Testing Dependencies
pytest>=8.4.0
pytest-asyncio>=1.0.0